// apps/admin/src/api/generated/schemas/emailTemplateRead.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * x-ear API
 * Backend API for x-ear CRM
 * OpenAPI spec version: 1.0.0
 */
import type { EmailTemplateReadCreatedAt } from './emailTemplateReadCreatedAt';
import type { EmailTemplateReadUpdatedAt } from './emailTemplateReadUpdatedAt';
import type { EmailTemplateReadDescription } from './emailTemplateReadDescription';
import type { EmailTemplateReadTitleTemplate } from './emailTemplateReadTitleTemplate';
import type { EmailTemplateReadBodyTemplate } from './emailTemplateReadBodyTemplate';
import type { EmailTemplateReadVariables } from './emailTemplateReadVariables';
import type { EmailTemplateReadEmailFromName } from './emailTemplateReadEmailFromName';
import type { EmailTemplateReadEmailFromAddress } from './emailTemplateReadEmailFromAddress';
import type { EmailTemplateReadEmailToType } from './emailTemplateReadEmailToType';
import type { EmailTemplateReadEmailToAddresses } from './emailTemplateReadEmailToAddresses';
import type { EmailTemplateReadEmailSubject } from './emailTemplateReadEmailSubject';
import type { EmailTemplateReadEmailBodyHtml } from './emailTemplateReadEmailBodyHtml';
import type { EmailTemplateReadEmailBodyText } from './emailTemplateReadEmailBodyText';
import type { EmailTemplateReadTriggerEvent } from './emailTemplateReadTriggerEvent';
import type { EmailTemplateReadTriggerConditions } from './emailTemplateReadTriggerConditions';
import type { EmailTemplateReadTemplateCategory } from './emailTemplateReadTemplateCategory';

/**
 * Schema for reading NotificationTemplate
 */
export interface EmailTemplateRead {
  /** Creation timestamp */
  createdAt?: EmailTemplateReadCreatedAt;
  /** Last update timestamp */
  updatedAt?: EmailTemplateReadUpdatedAt;
  /** Unique identifier for the resource */
  id: string;
  name: string;
  description?: EmailTemplateReadDescription;
  titleTemplate?: EmailTemplateReadTitleTemplate;
  bodyTemplate?: EmailTemplateReadBodyTemplate;
  channel?: string;
  variables?: EmailTemplateReadVariables;
  isActive?: boolean;
  emailFromName?: EmailTemplateReadEmailFromName;
  emailFromAddress?: EmailTemplateReadEmailFromAddress;
  emailToType?: EmailTemplateReadEmailToType;
  emailToAddresses?: EmailTemplateReadEmailToAddresses;
  emailSubject?: EmailTemplateReadEmailSubject;
  emailBodyHtml?: EmailTemplateReadEmailBodyHtml;
  emailBodyText?: EmailTemplateReadEmailBodyText;
  triggerEvent?: EmailTemplateReadTriggerEvent;
  triggerConditions?: EmailTemplateReadTriggerConditions;
  templateCategory?: EmailTemplateReadTemplateCategory;
}