// apps/admin/src/api/generated/schemas/notificationSend.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * x-ear API
 * Backend API for x-ear CRM
 * OpenAPI spec version: 1.0.0
 */
import type { NotificationSendTargetId } from './notificationSendTargetId';
import type { NotificationSendType } from './notificationSendType';
import type { NotificationSendChannel } from './notificationSendChannel';

export interface NotificationSend {
  targetType: string;
  targetId?: NotificationSendTargetId;
  title: string;
  message: string;
  type?: NotificationSendType;
  channel?: NotificationSendChannel;
}