// apps/admin/src/api/generated/schemas/responseEnvelopeListEmailTemplateRead.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * x-ear API
 * Backend API for x-ear CRM
 * OpenAPI spec version: 1.0.0
 */
import type { ResponseEnvelopeListEmailTemplateReadData } from './responseEnvelopeListEmailTemplateReadData';
import type { ResponseEnvelopeListEmailTemplateReadMessage } from './responseEnvelopeListEmailTemplateReadMessage';
import type { ResponseEnvelopeListEmailTemplateReadError } from './responseEnvelopeListEmailTemplateReadError';
import type { ResponseEnvelopeListEmailTemplateReadMeta } from './responseEnvelopeListEmailTemplateReadMeta';
import type { ResponseEnvelopeListEmailTemplateReadRequestId } from './responseEnvelopeListEmailTemplateReadRequestId';

export interface ResponseEnvelopeListEmailTemplateRead {
  success?: boolean;
  data?: ResponseEnvelopeListEmailTemplateReadData;
  message?: ResponseEnvelopeListEmailTemplateReadMessage;
  error?: ResponseEnvelopeListEmailTemplateReadError;
  meta?: ResponseEnvelopeListEmailTemplateReadMeta;
  requestId?: ResponseEnvelopeListEmailTemplateReadRequestId;
  timestamp?: string;
}